api/requirements.txt
kubernetes>=29.0.0
jsonschema>=4.21.1
fastapi[all]>=0.110.0
uvicorn>=0.27.1
cryptography>=42.0.5
pydantic>=2.6.3
pydantic-extra-types>=2.6.0
pydantic-settings>=2.2.1
pydantic_core>=2.16.3
PyJWT[crypto]>=2.8.0
asyncio>=3.4.3